Adding compression functions

+++ archivelib-rs-tdd/src/consts.rs
@@ -18,10 +18,10 @@ pub const CONST_N148_IS_4096: usize = 4096;
 pub const CONST_N149_IS_256: usize = 256;
 pub const CONST_N153_IS_4096: usize = 4096;
 
-pub const CONST_N142_IS_15: u16 = (MAX_COMPRESSION_FACTOR as u16 + 1);
-pub const CONST_N143_IS_9: u16 = 9;
-pub const CONST_N147_IS_5: u16 = 5;
-pub const CONST_N154_IS_4: u16 = 4;
-pub const CONST_N155_IS_8192: u16 = 8192;
-pub const CONST_N540_IS_5: u16 = 5;
+pub const CONST_N142_IS_15: usize = MAX_COMPRESSION_FACTOR as usize + 1;
+pub const CONST_N143_IS_9: usize = 9;
+pub const CONST_N147_IS_5: usize = 5;
+pub const CONST_N154_IS_4: usize = 4;
+pub const CONST_N155_IS_8192: usize = 8192;
+pub const CONST_N540_IS_5: usize = 5;
 // #define TRUE157 (-1)
