+++ archivelib-rs-tdd/src/expand/calculate_run_offset.rs
@@ -6,7 +6,7 @@ impl<R: BitRead, W: BitwiseWrite> RExpandData<R, W> {
   pub fn calculate_run_offset(&mut self) -> Result<u16> {
     let mut run_length276 = self.dat_arr241[(self.bits182 >> 8) as usize];
     let mut _283 = (1 << 7) as u16;
-    while run_length276 >= CONST_N142_IS_15 {
+    while run_length276 >= CONST_N142_IS_15 as u16 {
       if 0 != self.bits182 & _283 {
         run_length276 = self.dat_arr190[run_length276 as usize];
       } else {
